pyproject.toml
[project]
name = "crules"
version = "0.2.1"
description = "Cursor Rules Generator Script"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
keywords = ["cursor", "ai", "rules", "development"]
authors = [
    {name = "draeician", email = "draeician@gmail.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0.0",
    "setuptools>=42.0.0",
]

[project.urls]
Homepage = "https://github.com/draeician/crules"
Repository = "https://github.com/draeician/crules.git"
Documentation = "https://github.com/draeician/crules#readme"

[project.scripts]
crules = "crules.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["crules", "crules.rules"]

[tool.setuptools.package-data]
"crules.rules" = [
    "cursor.*",
    "default_cursorrules"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py39" 
